+++ src/connections/index.js
@@ -2,3 +2,4 @@ export { ConnectionInterface } from "./ConnectionInterface.js";
 export { BluetoothConnection } from "./BluetoothConnection.js";
 export { WebUSBConnection } from "./WebUSBConnection.js";
 export { WebSerialConnection } from "./WebSerialConnection.js";
+export { NetworkConnection } from "./NetworkConnection.js";